+++ lib/core/i18n/app_strings.dart
@@ -29,6 +29,7 @@ abstract interface class AppStrings {
   String get leaveLoyalty;
   String get chooseTablePrompt;
   String get chooseTable;
+  String get scanTable;
   String get callWaiter;
   String get bringBill;
   String get cart;